src/autoscaler/requirements.txt
kubernetes==18.20.0
google-auth
google-auth-httplib2
google-api-python-client
